# api/src/openapi.yml
openapi: 3.0.3
info:
  title: Dawson POS API
  version: 1.0.0
  description: >
    API para sistema de ventas, stock, caja, clientes, proveedores, compras, combos dinámicos,
    con control estricto por roles (ADMIN / CAJERO), dinero en centavos (integer) y combos por coincidencia exacta.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /products/{productId}:
    patch:
      tags: [Products]
      summary: (ADMIN) Actualizar presentación (incluye activo/inactivo)
      description: Solo ADMIN. Control manual de activo/inactivo.
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProductRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductAdminView" }
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /product-bases/{productBaseId}:
    patch:
      tags: [Products]
      summary: (ADMIN) Actualizar producto base (incluye activo/inactivo)
      parameters:
        - $ref: "#/components/parameters/ProductBaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProductBaseRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductBaseAdminView" }
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /product-bases:
    get:
      tags: [Products]
      summary: (ADMIN) Listar productos base (con stock)
      description: Solo ADMIN. Incluye stockActual y estado activo/inactivo.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductBaseListResponse" }
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    post:
      tags: [Products]
      summary: (ADMIN) Crear producto base
      description: Solo ADMIN.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProductBaseRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductBaseAdminView" }
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /products:
    get:
      tags: [Products]
      summary: (ADMIN) Listar presentaciones
      description: Solo ADMIN. Devuelve presentaciones aunque estén inactivas.
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductListResponse" }
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    post:
      tags: [Products]
      summary: (ADMIN) Crear presentación
      description: Solo ADMIN.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProductRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductAdminView" }
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /product-bases/{productBaseId}/stock-adjust:
    post:
      tags: [Products]
      summary: (ADMIN) Ajustar stock de un producto base
      description: >
        Solo ADMIN. Permite sumar o restar stock (decimales).
        Genera movimiento de stock (AJUSTE) y auditoría.
      parameters:
        - $ref: "#/components/parameters/ProductBaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StockAdjustRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductBaseAdminView" }
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      in: query
      name: pageSize
      schema: { type: integer, minimum: 1, maximum: 200, default: 25 }
...
    ProductId:
      in: path
      name: productId
      required: true
      schema: { type: string }
    ProductBaseId:
      in: path
      name: productBaseId
      required: true
      schema: { type: string }
  responses:
    UnauthorizedError:
      description: No autenticado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: unauthorized
            message: Token inválido o ausente
    ForbiddenError:
      description: Sin permisos (rol)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: forbidden
            message: No tiene permisos para esta acción
    NotFoundError:
      description: No encontrado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: not_found
            message: Recurso no encontrado
...
  schemas:
    UnidadStock:
      type: string
      enum: [LITROS, KILOS, UNIDADES]
    CreateProductBaseRequest:
      type: object
      required: [nombre, categoria, unidadStock]
      properties:
        nombre: { type: string, minLength: 1 }
        categoria: { $ref: "#/components/schemas/Categoria" }
        unidadStock: { $ref: "#/components/schemas/UnidadStock" }
        stockMinimo: { type: number, default: 0 }
        activo: { type: boolean, default: true }
    CreateProductRequest:
      type: object
      required: [productBaseId, nombre, categoria, tipo, cantidadDescuento, precioVenta, activo]
      properties:
        productBaseId: { type: string }
        nombre: { type: string, minLength: 1 }
        categoria: { $ref: "#/components/schemas/Categoria" }
        tipo: { $ref: "#/components/schemas/ProductoTipo" }
        cantidadDescuento: { type: number }
        precioVenta: { type: integer, description: Centavos }
        activo: { type: boolean }
    StockAdjustRequest:
      type: object
      required: [cantidad, motivo]
      properties:
        cantidad:
          type: number
          description: Puede ser positiva (sumar) o negativa (restar)
        motivo:
          type: string
          minLength: 1
    ProductBaseListItem:
      allOf:
        - $ref: "#/components/schemas/ProductBaseAdminView"
    ProductBaseListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ProductBaseListItem" }
    ProductListItem:
      allOf:
        - $ref: "#/components/schemas/ProductAdminView"
    ProductListResponse:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ProductListItem" }
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
...
    Categoria:
      type: string
      enum: [JABON_ROPA, SUAVIZANTE, DETERGENTE, LAVANDINA, DESODORANTE_PISO, OTRO]
    ProductoTipo:
      type: string
      enum: [GRANEL, UNITARIO]
...
    UpdateProductRequest:
      type: object
      properties:
        nombre: { type: string }
        precioVenta: { type: integer, description: Centavos }
        activo: { type: boolean }
    ProductAdminView:
      type: object
      required: [productId, productBaseId, nombre, categoria, tipo, cantidadDescuento, precioVenta, activo]
      properties:
        productId: { type: string }
        productBaseId: { type: string }
        nombre: { type: string }
        categoria: { $ref: "#/components/schemas/Categoria" }
        tipo: { $ref: "#/components/schemas/ProductoTipo" }
        cantidadDescuento: { type: number }
        precioVenta: { type: integer, description: Centavos }
        activo: { type: boolean }
    UpdateProductBaseRequest:
      type: object
      properties:
        nombre: { type: string }
        categoria: { $ref: "#/components/schemas/Categoria" }
        stockMinimo: { type: number }
        activo: { type: boolean }
    ProductBaseAdminView:
      type: object
      required: [productBaseId, nombre, categoria, unidadStock, stockActual, stockMinimo, activo]
      properties:
        productBaseId: { type: string }
        nombre: { type: string }
        categoria: { $ref: "#/components/schemas/Categoria" }
        unidadStock: { type: string, enum: [LITROS, KILOS, UNIDADES] }
        stockActual: { type: number }
        stockMinimo: { type: number }
        activo: { type: boolean }
    Error:
      type: object
      required: [error, message]
      properties:
        error: { type: string }
        message: { type: string }
...